requirements.txt
fastapi
uvicorn
gunicorn
openai
python-docx
PyMuPDF